requirements.txt
-r src/requirements.txt
python-dotenv
# Packages needed to call AOAI API with Azure Active Directory authentication
azure-identity
# Packages needed to call AOAI API with the OpenAI Python API
openai
# Packages needed to run notebook samples
jupyter